✨ update: 벤으로 인한 환불은 낙찰을 제외하고 환불, (이전 commit)벤 유저는 입찰 불가

+++ src/main/java/com/sparta/itsmine/domain/kakaopay/service/KakaoPayService.java
@@ -73,6 +73,10 @@ public class KakaoPayService {
 
 
     public KakaoPayReadyResponseDto ready(Long productId, User user, AuctionRequestDto requestDto) {
+
+        //ban 상태의 유저는 입찰 못하게 차단
+        user.checkBlock();
+
         // Request header
         HttpHeaders headers = new HttpHeaders();
         headers.add("Authorization", "DEV_SECRET_KEY " + kakaopaySecretKey);
@@ -274,6 +278,7 @@ public class KakaoPayService {
 
     public void deleteWithOutSuccessfulAuction(Long productId) {
         List<Auction> auctions = auctionRepository.findAllByProductIdWithOutMaxPrice(productId);
+        //왜냐하면 맥스값이 곧 낙찰가니까
 
         for (Auction auction : auctions) {
             if (auction.getStatus().equals(BID)) {
@@ -296,7 +301,7 @@ public class KakaoPayService {
                 KakaoPayTid kakaoPayTid = kakaoPayRepository.findByAuctionId(auction.getId());
                 kakaoCancel(kakaoPayTid.getTid());
                 auctionRepository.delete(auction);
-            } else {
+            } else if(auction.getStatus().equals(NEED_PAY)) {
                 auctionRepository.delete(auction);
             }
         }
